refactor(wp): move getRelatedChartsForVariable to Chart
