+++ baker/GrapherBaker.tsx
@@ -15,7 +15,6 @@ import {
     partition,
 } from "@ourworldindata/utils"
 import {
-    getRelatedChartsForVariable,
     getRelatedResearchAndWritingForVariable,
     isWordpressAPIEnabled,
     isWordpressDBEnabled,
@@ -64,6 +63,7 @@ import { GdocPost } from "../db/model/Gdoc/GdocPost.js"
 import { getShortPageCitation } from "../site/gdocs/utils.js"
 import { getSlugForTopicTag, getTagToSlugMap } from "./GrapherBakingUtils.js"
 import pMap from "p-map"
+import { getRelatedChartsForVariable } from "../db/model/Chart.js"
 
 const renderDatapageIfApplicable = async (
     grapher: GrapherInterface,
